+++ osidb/__init__.py
@@ -2,5 +2,5 @@
 osidb version
 """
 
-__version__ = "5.9.0"
+__version__ = "5.10.3"
 VERSION = __version__